added modified file that takes s3 url
